requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
tensorflow>=2.17.0
pillow>=10.0.0
firebase-admin>=6.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.21.0
aiohttp>=3.9.0
pydantic>=2.0.0
pydantic-settings>=2.9.0
requests>=2.31.0
